backend/terraform/requirements.txt
google-cloud-firestore==2.20.0
google-cloud-storage==3.0.0
